# IISNotionSearch/IISNotionSearch.API/Controllers/NotionController.cs
using IISNotionSearch.API.Abstractions.Controllers;
using Microsoft.AspNetCore.Mvc;

namespace IISNotionSearch.API.Controllers;

public class NotionController : BaseController
{
    // public NotionController()
    // {
    //     
    // }
    //
    // [HttpPost("search")]
    // public Task<IActionResult> Search([FromQuery] string query)
    // {
    //     
    // }
    //
    // [HttpGet("pages/{id}")]
    // public Task<IActionResult> GetPage([FromRoute] string id)
    // {
    //     
    // }
    //
    // [HttpPost("pages")]
    // public Task<IActionResult> CreatePage([FromBody] object pageData)
    // {
    //     
    // }
    //
    // [HttpPatch("pages/{id}")]
    // public Task<IActionResult> UpdatePage([FromRoute] string id, [FromBody] object pageData)
    // {
    //     
    // }
    //
    // [HttpDelete("pages/{id}")]
    // public Task<IActionResult> DeletePage([FromRoute] string id)
    // {
    //     
    // }
}