# IISNotionSearch/IISNotionSearch.Infrastructure/OpenApi/notion-pages-update-openapi.yaml
openapi: 3.1.0
info:
  title: Notion API
  version: 1.0.0
  termsOfService: >-
    https://notion.notion.site/Terms-and-Privacy-28ffdd083dc3473e9c2da6ec011b58ac
servers:
  - url: https://api.notion.com
security:
  - bearerAuth: []
...
paths:
  /v1/pages/{page_id}:
    patch:
      tags:
        - Pages
      summary: Update page
      operationId: patch-page
      parameters:
        - name: page_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/idRequest'
        - $ref: '#/components/parameters/notionVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                properties:
                  type: object
                  additionalProperties:
                    anyOf:
                      - title: Title
                        type: object
                        properties:
                          title:
                            type: array
                            items:
                              $ref: '#/components/schemas/richTextItemRequest'
                            maxItems: 100
                          type:
                            type: string
                            const: title
                        additionalProperties: false
                        required:
                          - title
                      - title: Rich Text
                        type: object
                        properties:
                          rich_text:
                            type: array
                            items:
                              $ref: '#/components/schemas/richTextItemRequest'
                            maxItems: 100
                          type:
                            type: string
                            const: rich_text
                        additionalProperties: false
                        required:
                          - rich_text
                      - title: Number
                        type: object
                        properties:
                          number:
                            type:
                              - number
                              - 'null'
                          type:
                            type: string
                            const: number
                        additionalProperties: false
                        required:
                          - number
                      - title: Url
                        type: object
                        properties:
                          url:
                            anyOf:
                              - $ref: '#/components/schemas/textRequest'
                              - type: 'null'
                          type:
                            type: string
                            const: url
                        additionalProperties: false
                        required:
                          - url
                      - title: Select
                        type: object
                        properties:
                          select:
                            anyOf:
                              - type: object
                                properties:
                                  id:
                                    $ref: '#/components/schemas/stringRequest'
                                  name:
                                    $ref: '#/components/schemas/textRequest'
                                  color:
                                    $ref: '#/components/schemas/selectColor'
                                  description:
                                    anyOf:
                                      - $ref: '#/components/schemas/textRequest'
                                      - type: 'null'
                                additionalProperties: false
                                required:
                                  - id
                              - type: object
                                properties:
                                  name:
                                    $ref: '#/components/schemas/textRequest'
                                  id:
                                    $ref: '#/components/schemas/stringRequest'
                                  color:
                                    $ref: '#/components/schemas/selectColor'
                                  description:
                                    anyOf:
                                      - $ref: '#/components/schemas/textRequest'
                                      - type: 'null'
                                additionalProperties: false
                                required:
                                  - name
                              - type: 'null'
                          type:
                            type: string
                            const: select
                        additionalProperties: false
                        required:
                          - select
                      - title: Multi Select
                        type: object
                        properties:
                          multi_select:
                            type: array
                            items:
                              anyOf:
                                - type: object
                                  properties:
                                    id:
                                      $ref: '#/components/schemas/stringRequest'
                                    name:
                                      $ref: '#/components/schemas/textRequest'
                                    color:
                                      $ref: '#/components/schemas/selectColor'
                                    description:
                                      anyOf:
                                        - $ref: '#/components/schemas/textRequest'
                                        - type: 'null'
                                  additionalProperties: false
                                  required:
                                    - id
                                - type: object
                                  properties:
                                    name:
                                      $ref: '#/components/schemas/textRequest'
                                    id:
                                      $ref: '#/components/schemas/stringRequest'
                                    color:
                                      $ref: '#/components/schemas/selectColor'
                                    description:
                                      anyOf:
                                        - $ref: '#/components/schemas/textRequest'
                                        - type: 'null'
                                  additionalProperties: false
                                  required:
                                    - name
                            maxItems: 100
                          type:
                            type: string
                            const: multi_select
                        additionalProperties: false
                        required:
                          - multi_select
                      - title: People
                        type: object
                        properties:
                          people:
                            type: array
                            items:
                              anyOf:
                                - $ref: >-
                                    #/components/schemas/partialUserObjectRequest
                                - $ref: '#/components/schemas/groupObjectRequest'
                            maxItems: 100
                          type:
                            type: string
                            const: people
                        additionalProperties: false
                        required:
                          - people
                      - title: Email
                        type: object
                        properties:
                          email:
                            anyOf:
                              - $ref: '#/components/schemas/stringRequest'
                              - type: 'null'
                          type:
                            type: string
                            const: email
                        additionalProperties: false
                        required:
                          - email
                      - title: Phone Number
                        type: object
                        properties:
                          phone_number:
                            anyOf:
                              - $ref: '#/components/schemas/stringRequest'
                              - type: 'null'
                          type:
                            type: string
                            const: phone_number
                        additionalProperties: false
                        required:
                          - phone_number
                      - title: Date
                        type: object
                        properties:
                          date:
                            anyOf:
                              - $ref: '#/components/schemas/dateRequest'
                              - type: 'null'
                          type:
                            type: string
                            const: date
                        additionalProperties: false
                        required:
                          - date
                      - title: Checkbox
                        type: object
                        properties:
                          checkbox:
                            type: boolean
                          type:
                            type: string
                            const: checkbox
                        additionalProperties: false
                        required:
                          - checkbox
                      - title: Relation
                        type: object
                        properties:
                          relation:
                            type: array
                            items:
                              $ref: >-
                                #/components/schemas/relationItemPropertyValueResponse
                            maxItems: 100
                          type:
                            type: string
                            const: relation
                        additionalProperties: false
                        required:
                          - relation
                      - title: Files
                        type: object
                        properties:
                          files:
                            type: array
                            items:
                              anyOf:
                                - $ref: >-
                                    #/components/schemas/internalOrExternalFileWithNameRequest
                                - $ref: >-
                                    #/components/schemas/fileUploadWithOptionalNameRequest
                            maxItems: 100
                          type:
                            type: string
                            const: files
                        additionalProperties: false
                        required:
                          - files
                      - title: Status
                        type: object
                        properties:
                          status:
                            anyOf:
                              - type: object
                                properties:
                                  id:
                                    $ref: '#/components/schemas/stringRequest'
                                  name:
                                    $ref: '#/components/schemas/textRequest'
                                  color:
                                    $ref: '#/components/schemas/selectColor'
                                  description:
                                    anyOf:
                                      - $ref: '#/components/schemas/textRequest'
                                      - type: 'null'
                                additionalProperties: false
                                required:
                                  - id
                              - type: object
                                properties:
                                  name:
                                    $ref: '#/components/schemas/textRequest'
                                  id:
                                    $ref: '#/components/schemas/stringRequest'
                                  color:
                                    $ref: '#/components/schemas/selectColor'
                                  description:
                                    anyOf:
                                      - $ref: '#/components/schemas/textRequest'
                                      - type: 'null'
                                additionalProperties: false
                                required:
                                  - name
                              - type: 'null'
                          type:
                            type: string
                            const: status
                        additionalProperties: false
                        required:
                          - status
                      - title: Place
                        type: object
                        properties:
                          place:
                            type:
                              - object
                              - 'null'
                            properties:
                              lat:
                                type: number
                              lon:
                                type: number
                              name:
                                type:
                                  - string
                                  - 'null'
                              address:
                                type:
                                  - string
                                  - 'null'
                              aws_place_id:
                                type:
                                  - string
                                  - 'null'
                              google_place_id:
                                type:
                                  - string
                                  - 'null'
                            required:
                              - lat
                              - lon
                          type:
                            type: string
                            const: place
                        additionalProperties: false
                        required:
                          - place
                      - title: Verification
                        type: object
                        properties:
                          verification:
                            anyOf:
                              - type: object
                                properties:
                                  state:
                                    type: string
                                    const: verified
                                  date:
                                    $ref: '#/components/schemas/dateRequest'
                                required:
                                  - state
                              - type: object
                                properties:
                                  state:
                                    type: string
                                    const: unverified
                                required:
                                  - state
                          type:
                            type: string
                            const: verification
                        additionalProperties: false
                        required:
                          - verification
                icon:
                  anyOf:
                    - $ref: '#/components/schemas/pageIconRequest'
                    - type: 'null'
                cover:
                  anyOf:
                    - $ref: '#/components/schemas/pageCoverRequest'
                    - type: 'null'
                is_locked:
                  description: >-
                    Whether the page should be locked from editing in the Notion
                    app UI. If not provided, the locked state will not be
                    updated.
                  type: boolean
                template:
                  anyOf:
                    - type: object
                      properties:
                        type:
                          type: string
                          const: default
                        timezone:
                          $ref: '#/components/schemas/templateTimezone'
                      required:
                        - type
                    - type: object
                      properties:
                        type:
                          type: string
                          const: template_id
                        template_id:
                          $ref: '#/components/schemas/idRequest'
                        timezone:
                          $ref: '#/components/schemas/templateTimezone'
                      required:
                        - type
                        - template_id
                erase_content:
                  description: >-
                    Whether to erase all existing content from the page. When
                    used with a template, the template content replaces the
                    existing content. When used without a template, simply
                    clears the page content.
                  type: boolean
                in_trash:
                  type: boolean
                is_archived:
                  type: boolean
              additionalProperties: false
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/pageObjectResponse'
                  - $ref: '#/components/schemas/partialPageObjectResponse'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_api_400'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_api_401'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_api_403'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_api_404'
        '409':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_api_409'
        '429':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_api_429'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_api_500'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_api_503'
        '504':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_api_504'
      x-codeSamples:
        - lang: javascript
          label: TypeScript SDK
          source: |-
            import { Client } from "@notionhq/client"

            const notion = new Client({ auth: process.env.NOTION_API_KEY })

            const response = await notion.pages.update({
              page_id: "b55c9c91-384d-452b-81db-d1ef79372b75",
              properties: {
                Name: {
                  title: [{ text: { content: "Updated Title" } }]
                }
              }
            })
components:
  schemas:
    idRequest:
      type: string
    richTextItemRequest:
      allOf:
        - $ref: '#/components/schemas/richTextItemRequestCommon'
        - oneOf:
            - $ref: '#/components/schemas/textRichTextItemRequest'
            - $ref: '#/components/schemas/mentionRichTextItemRequest'
            - $ref: '#/components/schemas/equationRichTextItemRequest'
    textRequest:
      type: string
      maxLength: 2000
      minLength: 1
    stringRequest:
      type: string
      maxLength: 100
      minLength: 1
    selectColor:
      type: string
      enum:
        - default
        - gray
        - brown
        - orange
        - yellow
        - green
        - blue
        - purple
        - pink
        - red
      description: >-
        One of: `default`, `gray`, `brown`, `orange`, `yellow`, `green`, `blue`,
        `purple`, `pink`, `red`
    partialUserObjectRequest:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/idRequest'
          description: The ID of the user.
        object:
          type: string
          const: user
          description: The user object type name.
      required:
        - id
    groupObjectRequest:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/idRequest'
        name:
          type:
            - string
            - 'null'
        object:
          type: string
          const: group
      additionalProperties: false
      required:
        - id
    dateRequest:
      type: object
      properties:
        start:
          type: string
          format: date
          description: The start date of the date object.
        end:
          oneOf:
            - type: string
              format: date
            - type: 'null'
          description: The end date of the date object, if any.
        time_zone:
          oneOf:
            - $ref: '#/components/schemas/timeZoneRequest'
            - type: 'null'
          description: >-
            The time zone of the date object, if any. E.g. America/Los_Angeles,
            Europe/London, etc.
      additionalProperties: false
      required:
        - start
    relationItemPropertyValueResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/idRequest'
      required:
        - id
    internalOrExternalFileWithNameRequest:
      anyOf:
        - title: File
          type: object
          properties:
            file:
              $ref: '#/components/schemas/internalFileRequest'
            name:
              $ref: '#/components/schemas/stringRequest'
            type:
              type: string
              const: file
          additionalProperties: false
          required:
            - file
            - name
        - title: External
          type: object
          properties:
            external:
              $ref: '#/components/schemas/externalFileRequest'
            name:
              $ref: '#/components/schemas/stringRequest'
            type:
              type: string
              const: external
          additionalProperties: false
          required:
            - external
            - name
    fileUploadWithOptionalNameRequest:
      type: object
      properties:
        file_upload:
          $ref: '#/components/schemas/fileUploadIdRequest'
        type:
          type: string
          const: file_upload
        name:
          $ref: '#/components/schemas/stringRequest'
      additionalProperties: false
      required:
        - file_upload
    pageIconRequest:
      oneOf:
        - $ref: '#/components/schemas/fileUploadPageIconRequest'
        - $ref: '#/components/schemas/emojiPageIconRequest'
        - $ref: '#/components/schemas/externalPageIconRequest'
        - $ref: '#/components/schemas/customEmojiPageIconRequest'
        - $ref: '#/components/schemas/iconPageIconRequest'
    pageCoverRequest:
      oneOf:
        - $ref: '#/components/schemas/fileUploadPageCoverRequest'
        - $ref: '#/components/schemas/externalPageCoverRequest'
    templateTimezone:
      description: >-
        IANA timezone to use when resolving template variables like @now and
        @today (e.g. 'America/New_York'). Defaults to the authorizing user's
        timezone for public integrations, or UTC for internal integrations.
      examples:
        - America/New_York
        - Europe/London
        - Asia/Tokyo
      type: string
    pageObjectResponse:
      type: object
      properties:
        object:
          type: string
          const: page
          description: The page object type name.
        id:
          $ref: '#/components/schemas/idResponse'
          description: The ID of the page.
        created_time:
          type: string
          format: date-time
          description: Date and time when this page was created.
        last_edited_time:
          type: string
          format: date-time
          description: Date and time when this page was last edited.
        in_trash:
          type: boolean
          description: Whether the page is in trash.
        is_archived:
          type: boolean
          description: Whether the page has been archived.
        is_locked:
          type: boolean
          description: Whether the page is locked from editing in the Notion app UI.
        url:
          type: string
          description: The URL of the Notion page.
        public_url:
          oneOf:
            - type: string
            - type: 'null'
          description: >-
            The public URL of the Notion page, if it has been published to the
            web.
        parent:
          $ref: '#/components/schemas/parentForBlockBasedObjectResponse'
          description: Information about the page's parent.
        properties:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/pagePropertyValueWithIdResponse'
          description: Property values of this page.
        icon:
          oneOf:
            - $ref: '#/components/schemas/pageIconResponse'
            - type: 'null'
          description: Page icon.
        cover:
          oneOf:
            - $ref: '#/components/schemas/pageCoverResponse'
            - type: 'null'
          description: Page cover image.
        created_by:
          $ref: '#/components/schemas/partialUserObjectResponse'
          description: User who created the page.
        last_edited_by:
          $ref: '#/components/schemas/partialUserObjectResponse'
          description: User who last edited the page.
      additionalProperties: false
      required:
        - object
        - id
        - created_time
        - last_edited_time
        - in_trash
        - is_archived
        - is_locked
        - url
        - public_url
        - parent
        - properties
        - icon
        - cover
        - created_by
        - last_edited_by
    partialPageObjectResponse:
      type: object
      properties:
        object:
          type: string
          const: page
          description: The page object type name.
        id:
          $ref: '#/components/schemas/idResponse'
          description: The ID of the page.
      additionalProperties: false
      required:
        - object
        - id
    error_api_400:
      allOf:
        - $ref: '#/components/schemas/publicApiCommonErrorResponse'
        - type: object
          properties:
            code:
              enum:
                - invalid_json
                - invalid_request_url
                - invalid_request
                - missing_version
                - validation_error
            status:
              const: 400
          required:
            - code
            - status
          additionalProperties: false
    error_api_401:
      allOf:
        - $ref: '#/components/schemas/publicApiCommonErrorResponse'
        - type: object
          properties:
            code:
              enum:
                - unauthorized
            status:
              const: 401
          required:
            - code
            - status
          additionalProperties: false
    error_api_403:
      allOf:
        - $ref: '#/components/schemas/publicApiCommonErrorResponse'
        - type: object
          properties:
            code:
              enum:
                - restricted_resource
            status:
              const: 403
          required:
            - code
            - status
          additionalProperties: false
    error_api_404:
      allOf:
        - $ref: '#/components/schemas/publicApiCommonErrorResponse'
        - type: object
          properties:
            code:
              enum:
                - object_not_found
            status:
              const: 404
          required:
            - code
            - status
          additionalProperties: false
    error_api_409:
      allOf:
        - $ref: '#/components/schemas/publicApiCommonErrorResponse'
        - type: object
          properties:
            code:
              enum:
                - conflict_error
            status:
              const: 409
          required:
            - code
            - status
          additionalProperties: false
    error_api_429:
      allOf:
        - $ref: '#/components/schemas/publicApiCommonErrorResponse'
        - type: object
          properties:
            code:
              enum:
                - rate_limited
            status:
              const: 429
          required:
            - code
            - status
          additionalProperties: false
    error_api_500:
      allOf:
        - $ref: '#/components/schemas/publicApiCommonErrorResponse'
        - type: object
          properties:
            code:
              enum:
                - internal_server_error
            status:
              const: 500
          required:
            - code
            - status
          additionalProperties: false
    error_api_503:
      allOf:
        - $ref: '#/components/schemas/publicApiCommonErrorResponse'
        - type: object
          properties:
            code:
              enum:
                - service_unavailable
            status:
              const: 503
          required:
            - code
            - status
          additionalProperties: false
    error_api_504:
      allOf:
        - $ref: '#/components/schemas/publicApiCommonErrorResponse'
        - type: object
          properties:
            code:
              enum:
                - gateway_timeout
            status:
              const: 504
          required:
            - code
            - status
          additionalProperties: false
    richTextItemRequestCommon:
      type: object
      properties:
        annotations:
          $ref: '#/components/schemas/annotationRequest'
          description: >-
            All rich text objects contain an annotations object that sets the
            styling for the rich text.
    textRichTextItemRequest:
      type: object
      properties:
        type:
          type: string
          const: text
          description: Always `text`
        text:
          type: object
          properties:
            content:
              type: string
              maxLength: 2000
              description: The actual text content of the text.
            link:
              oneOf:
                - type: object
                  properties:
                    url:
                      type: string
                      examples:
                        - https://www.notion.com
                      description: The URL of the link.
                  required:
                    - url
                - type: 'null'
              description: >-
                An object with information about any inline link in this text,
                if included.
          additionalProperties: false
          required:
            - content
          description: >-
            If a rich text object's type value is `text`, then the corresponding
            text field contains an object including the text content and any
            inline link.
      required:
        - text
      title: Text
...
    equationRichTextItemRequest:
      type: object
      properties:
        type:
          type: string
          const: equation
          description: Always `equation`
        equation:
          type: object
          properties:
            expression:
              type: string
              examples:
                - e=mc^2
              description: A KaTeX compatible string.
          required:
            - expression
          description: >-
            Notion supports inline LaTeX equations as rich text objects with a
            type value of `equation`.
      required:
        - equation
      title: Equation
    timeZoneRequest:
      type: string
    internalFileRequest:
      type: object
      properties:
        url:
          type: string
        expiry_time:
          type: string
          format: date-time
      additionalProperties: false
      required:
        - url
    externalFileRequest:
      type: object
      properties:
        url:
          $ref: '#/components/schemas/textRequest'
      additionalProperties: false
      required:
        - url
    fileUploadIdRequest:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/idRequest'
      additionalProperties: false
      required:
        - id
    fileUploadPageIconRequest:
      type: object
      properties:
        type:
          type: string
          const: file_upload
          description: Always `file_upload`
        file_upload:
          type: object
          properties:
            id:
              type: string
              description: ID of a FileUpload object that has the status `uploaded`.
          required:
            - id
      required:
        - file_upload
      title: File Upload
...
    idResponse:
      type: string
      format: uuid
...
  parameters:
    notionVersion:
      name: Notion-Version
      in: header
      required: true
      schema:
        enum:
          - '2026-03-11'
      description: >-
        The [API version](/reference/versioning) to use for this request. The
        latest version is `2026-03-11`.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer